Fixed incorrect PTX parsing of `ret` instruction after branch label (#17859)
The PTX parser replaces PTX code with inline PTX code (using inline ASM blocks).
It considers a branch label and the immediate instruction as a single unit to process.  
During the ASM->CUDA transform step,  it searches for the `ret` instruction in the string and replaces the whole statement and not the substring that contains the `ret;` instruction. which means an expression like:

```asm

BB0_1:
ret;
```

is parsed as: 

```asm 

BB0_1: ret;

```

and then transformed to: 

```asm

bra RETTGT

``` 

instead of:

```asm 

BB0_1: bra RETTGT

```

This merge request fixes this bug.

Authors:
  - Basit Ayantunde (https://github.com/lamarrr)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/17859