Get rid of the deprecated `thrust::identity` (#17942)
This PR removes the usage of the deprecated `thrust::identity` and replaces it with `cuda::std::identity` where applicable. In cases where `thrust::identity` was used as a type conversion callable and `cuda::std::identity` is not suitable, it uses a new utility called `cast_fn` in the `detail` namespace.

Authors:
  - Yunsong Wang (https://github.com/PointKernel)

Approvers:
  - Michael Schellenberger Costa (https://github.com/miscco)
  - David Wendt (https://github.com/davidwendt)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/17942