Add fmt to the test/benchmarks env (#18173)
With the transition to rapids-logger libcudf no longer has a direct
dependency on fmt. However, the benchmarks still have a transitive
dependency via nvbench. Interestingly, the google benchmarks also seem
to have a dependency there and I have not uncovered why yet. With the
merge of https://github.com/rapidsai/rapids-cmake/pull/790, builds using
rapids-cmake no longer require a patched version of fmt anywhere, which
is resulting in benchmark builds picking up fmt from the environment.
This results in overlinking issues for the benchmarks binaries, and if
the check is disabled the binaries fail at runtime since fmt will not be
present in the runtime environment. This PR temporarily patches this
issue by adding the dependency until we can figure out how we want to
deal with it going forward.