Make the version file inside cudf.pandas not a symlink (#18198)
The VERSION file is meant to be a single source of truth for the package version that we read everywhere. As a result, we also symlink it in a handful of places that need it, in particular in places where we build a Python package and want to read the version to embed it into the package metadata. Unfortunately, due to https://github.com/scikit-build/scikit-build-core/issues/647, this results in editable installs of our packages having an incorrect `__path__` attribute that includes the root directory of the repository (where the real VERSION file lives) in addition to the root directory of the package (where the symlink lives). In most cases this causes no real problems, but it is problematic for cudf.pandas in particular because [we use the `__path__` to set the denylist](https://github.com/rapidsai/cudf/blob/5ec865acbb0425b592296bb5dbdee5d9590b9a57/python/cudf/cudf/pandas/module_accelerator.py#L410). Since the root directory of the repository gets added to the denylist and that is where the tests live, tests using cudf.pandas do not work when run with editable installations.

Note that the same problem does not arise when a pytest tst is converted to a normal Python module and run with `python -m cudf.pandas ${TEST_FILE}` because in that case [the `calling_module` we get from the frame](https://github.com/rapidsai/cudf/blob/5ec865acbb0425b592296bb5dbdee5d9590b9a57/python/cudf/cudf/pandas/module_accelerator.py#L558) does not contain the full path to the file, only the raw filename. That is normally what you would expect as a user running an arbitrary script where the frames would only include full path to things outside of your working tree. The problem is that the main program when you run tests is pytest, and pytest has to then launch runs using the full paths to the test files.

We should revert this change whenever scikit-build/scikit-build-core#647 is resolved.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Murray (https://github.com/Matt711)
  - Bradley Dice (https://github.com/bdice)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/18198