xfail narwhals sqlframe tests (#18297)
The narwhals tests are failing due to a new Narwhals release (1.31) and polars release (1.25). The majority of the failures are in the polars pytest run. They will be fixed in #18132 which will update our polars version pinning to 1.25. 

The remaining tests which fail during the cudf and cudf.pandas pytest runs are skipped/xfailed in this PR. They fail because duckdb is not installed as a test dependency in Narwhals. Is it missing somewhere in https://github.com/narwhals-dev/narwhals/blob/main/pyproject.toml?

CC @MarcoGorelli

Authors:
  - Matthew Murray (https://github.com/Matt711)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/18297