Cache column view creation from arrow types (#18302)
The new types introduced in #18084 use the preexisting
`to_arrow_device*` functions to produce views. Due to not every cudf
type mapping perfectly to an Arrow type (historically decimals had
misalignment until recent versions of arrow, and we still store boolean
columns as-is rather than as bit columns like arrow) there are cases
where that conversion requires allocating new memory. Previously there
was no way to cache that conversion. Now, we can store the intermediate
in the new types to avoid needing to reallocate on repeated calls. This
change also synchronizes the APIs of the corresponding vanilla cudf
column/table types. To improve that synchronization, the view creation
happens upon creation of these types, allowing us to drop the stream and
mr parameters from the view methods. For readability, I also aligned the
ordering of the declarations in the header and the definitions in the
files.

I have not benchmarked the impact of these changes yet since we are not
using these APIs anywhere significant yet. I plan to add in benchmarks
as part of the PRs to leverage the new types from Python, at which point
I plan to optimize as needed.