Use the mapped buffer for all read operations in the memory-mapped source; switch default source to the kvikIO one (#18204)
Closes https://github.com/rapidsai/cudf/issues/17228

The current implementation of the memory-mapped source only uses the mapped buffer for host reads. Such reads make a tiny portion of the total data read for all readers, so the mapping is not really utilized.

This PR changes the behavior of this source to always use the mapped buffer.
Also changed the default to non-mapped source, as the old behavior of the mapped source was effectively much closer to non-mapped source and I want to keep the impact on the default behavior to the minimum.

Other changes: CSV reader used to rely on memory mapping for byte range validation. Added a check for the byte range start to the reader and adjusted a python test that depended on the exception type and message.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/18204