Fix alpha versions of cudf package. (#18429)
## Description
In PR #18198, we created a standalone file `python/cudf/cudf/VERSION`
that is not a symlink to `VERSION`. This was meant to fix issues with
editable installations. However, because of this, we also need to write
out the alpha version information to `python/cudf/cudf/VERSION` at build
time. This PR fixes that.

## Checklist
- [x] I am familiar with the [Contributing
Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.