Refcount-based dropping of cached evaluations in cudf-polars executor (#18430)
Previously we held on to evaluated `Cache` nodes for the lifetime of the query evaluation. This is more memory pressure than necessary. To fix this, track the number of times we see the node and drop the cached value once we see it for the last time.

So that the same thing occurs in the partitioned executor, implement specialised hashing and equality for cache nodes: their key is guaranteed to be unique by polars within a query, so that is sufficient for the hash key.

- Closes #17363

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Tom Augspurger (https://github.com/TomAugspurger)
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/18430