Start removal of vector factories with `_sync` suffix by deprecating them and adding versions without the suffix (#18414)
The `_sync` suffix is redundant because all functions that don't have the `_async` suffix should be assumed to be synchronous.

Deprecated the factory functions for now because they might be used outside of libcudf. Should be removed in a future release.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/18414