Fix overflow for `MERGE_M2` groupby aggregation (#18546)
This fixes an overflow issue for the `MERGE_M2` groupby aggregation that has been reported in https://github.com/NVIDIA/spark-rapids/issues/12468. In particular, the current implementation of this aggregation requires the `count` input to be a column of type `INT32`, which cannot handle groups with sizes larger than `INT_MAX` (2B rows). To fix that, `count` type is changed to a wider type, which can be either `INT64` or `FLOAT64`.

This is also a breaking change and requires the downstream applications (Spark-Rapids) to change the corresponding data structure.

Authors:
  - Nghia Truong (https://github.com/ttnghia)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)
  - MithunR (https://github.com/mythrocks)
  - Yunsong Wang (https://github.com/PointKernel)
  - Tianyu Liu (https://github.com/kingcrimsontianyu)

URL: https://github.com/rapidsai/cudf/pull/18546