Add APIs to check ORC and Parquet compression support at runtime (#18578)
Issue #18445

Add APIs that check compression support at runtime (env vars and such), to be able to cleanly skip tests when the tested compression type is not supported. This removed the direct use of the nvcomp adapter in ORC and parquet tests.

Also use the new APIs in the readers and writers to help keep the implementation up to date.

Other changes:

- Minor improvements to internal APIs that check compression support.
- Remove outdated checks for ZSTD support as it's now guaranteed.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Yunsong Wang (https://github.com/PointKernel)

Approvers:
  - Paul Mattione (https://github.com/pmattione-nvidia)
  - Shruti Shivakumar (https://github.com/shrshi)
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/18578