Pin to cython<3.1 (#18617)
A 3.1 release candidate changes the way C++ code is emitted for Cython declarations like:

    ctypedef some_type alias_name
    cdef class Foo:
        cdef alias_name whatever

Previously, we would get

    new((void*)&(p->whatever)) alias_name();

Now, instead, we get

    new((void*)&(p->whatever)) some_type();

This breaks the usage of a `ctypedef` with `std::variant` in the ORC metadata, since we end up with

    new((void*)&(p->whatever)) std::variant();

Rather than

    new((void*)&(p->whatever)) cudf::io::statistics_type();

The latter is well formed, it default-constructs a `std::variant` that is defined as `std::variant<std::monostate, ...>` in libcudf headers. The former is ill-formed in two ways:

1. A variant must contain at least one default-constructible type
2. Even if we could default-construct, we're now constructing the wrong type.

As a work-around, let us pin to Cython 3.0, until we can figure out the correct way to deal with this and in case there are other fall-outs.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - James Lamb (https://github.com/jameslamb)

URL: https://github.com/rapidsai/cudf/pull/18617