Fix device compression when writing Parquet files without using nvCOMP (#18644)
Fixed an issue in the Parquet writer where all compression would fail when device compression with internal kernels is used.
Root cause is that a compressed chunk size was not being set, as it's not required for the nvCOMP implementation.

Expanded a few tests to use internal kernels for compression; planning to make more extensive test changes to improve test coverage of compression/decompression internal kernels in a separate PR.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Shruti Shivakumar (https://github.com/shrshi)
  - Nghia Truong (https://github.com/ttnghia)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/18644