Fix overflow case and clean up some logic (#18734)
The calculation of the size in bytes for the CUDA Array Interface for pylibcudf Column objects produced from a column_view and an arbitrary owner previously ran the risk of overflow because the arithmetic was performed on int32 types but that is actually the maximum size in number of elements, not bytes. Since the CAI is a Python object, we can do the arithmetic with pure Python (infinite precision) integers to avoid this problem. In the process of fixing this bug, this PR also does some minor cleanup of the various cases handled in the size calculation.

Resolves #18598

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/18734