Add nvtext substring deduplication APIs (Part 2) (#18104)
Adds new set of nvtext substring deduplication APIs

```cpp
std::unique_ptr<cudf::column> resolve_duplicates(
  cudf::strings_column_view const& input,
  cudf::device_span<int32_t const> indices,
  int32_t min_width,
  rmm::cuda_stream_view stream,
  rmm::device_async_resource_ref mr);
```
Accepts an input column and suffix array and returns duplicates.

```cpp
std::unique_ptr<cudf::column> resolve_duplicates_pair(
  cudf::strings_column_view const& input1,
  cudf::device_span<int32_t const> indices1,
  cudf::strings_column_view const& input2,
  cudf::device_span<int32_t const> indices2,
  int32_t min_width,
  rmm::cuda_stream_view stream,
  rmm::device_async_resource_ref mr);
```
This returns duplicates found between two columns of strings given suffix arrays for both.

The python interfaces have the same names but drop the first parameter since they are called an existing strings column. Here is an example usage where the `duplicate_input` column is divided in half and both halves are used in a call to `resolve_duplicates_pair`

```python
    text = duplicate_input
    L = int(len(text)/2)
    text1 = text[0:L]
    text2 = text[L:]

    input1 = cudf.Series(text1)
    sa1 = input1.str.build_suffix_array(0)
    input2 = cudf.Series(text2)
    sa2 = input2.str.build_suffix_array(0)
    result = input1.str.resolve_duplicates_pair(sa1, input2, sa2, 50)
```
The result will be a strings column of duplicates found between `text1` and `text2`

The `build_suffix_array` returns int32 values per input byte. So the output is 4x the input size.
Additionally, the internal sort routine requires an extra 4x of temporary memory.
This means the memory footprint for `build_suffix_array` will be about 8x the size of the input.
Also, this function will be about 95% of the runtime and care should be taken to not to call this repeatedly on the same input data.

The `resolve_duplicates` returns the duplicates so the size of the output is ultimately based on the actual number of duplicates within the strings column. Additionally, the working memory is 2x the input size plus about 25x times the number of duplicates (not the byte size but the actual duplicate count).
And this function is only about 5% of the total runtime (combined with `build_suffix_array`).

Similarly the `resolve_duplicates_pair` has about the same memory footprint where the working memory is 2x the larger or the 2 inputs and still 25x times the number of duplicates.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)
  - Bradley Dice (https://github.com/bdice)
  - Basit Ayantunde (https://github.com/lamarrr)

URL: https://github.com/rapidsai/cudf/pull/18104