Temporarily revert "Refactor JNI error handling (#18983)" (#19076)
Currently the cuDF builds for 25.08 show a failed build for spark-rapids-jni due to a conflict between the spark-rapids-jni 25.06, which is being picked up due to it being the active branch in jni, built against cuDF 25.08, which has some breaking changes introduced in the commit I am reverting here.

Once we move to 25.08 consistently we'll be able to bring this work back. Please also check:

https://github.com/NVIDIA/spark-rapids-jni/pull/3412

Authors:
  - Alessandro Bellina (https://github.com/abellina)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/19076