Temporary workaround for incorrect ``SplitScan`` results in cuDF-Polars (#19071)
This is a temporary workaround for https://github.com/rapidsai/cudf/issues/19064

The streaming cudf-polars executor uses `skip_rows`/`n_rows` to target a specific domain of a large Parquet file within a single task. Due to a bug in libcudf, this approach can produce incorrect results when a predicate is also specified. This PR simply avoids passing the predicate down to libcudf, and applies the mask **after** the read. It should be more efficient to avoid IO, but we can't do that until the libcudf bug is fixed.

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/19071