Allow setting `StreamingExecutor.target_partition_size` with an environment variable (#19316)
This provides a way to set the default `target_partition_size` for the cudf-polars streaming executor with an environment variable. The default behavior is unchanged: use a fraction of the device size, as reported by pynvml, with a warning if pynvml can't be found. Setting it via the environment is useful for usage through libraries like Narwhals, where AFAICT there isn't a way to pass through arguments to the engine.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/19316