Upgrade `gcc-toolset` for Java/JNI build to version 14 (#19500)
As RAPIDS is moving to gcc-14, we should upgrade the rest of the build system (Java/JNI) to match it.

This only changes in the build and test for Java/JNI, no implementation change is needed.

Build for `spark-rapids-jni` will not pass until https://github.com/NVIDIA/spark-rapids-jni/pull/3568 is merged. But merging this should not be blocked by it.

Authors:
  - Nghia Truong (https://github.com/ttnghia)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Peixin (https://github.com/pxLi)

URL: https://github.com/rapidsai/cudf/pull/19500