Propagate exceptions thrown in async IO operations (#19628)
closes #19586

In some places in libcudf, `std::future::wait()` is called to wait on asynchronous IO operations. The issue with this is that wait does not propagate any exceptions that may be thrown.

This PR replaces the `wait()` calls with `get()` and adds readers and writers tests to make sure exceptions are re-thrown.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Amin Aramoon (https://github.com/aminaramoon)
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/19628