Support output-type for MEDIAN/QUANTILE aggregation in cudf::reduce (#19267)
Adds support to return the aggregation result for MEDIAN and QUANTILE as specified in the `output_type` parameter in the `cudf::reduce` API.
The underlying code calls `cudf::detail::quantile` which has an `exact` parameter that returns a `double` type if set to true. Otherwise, it will return a value using the input type.

The code change here is to set `exact` to true only if the `output_type` for `cudf::reduce` is set to `FLOAT64`
This means the current behavior can be achieved by always setting the `output_type` to `FLOAT64`.

Authors:
  - David Wendt (https://github.com/davidwendt)
  - Matthew Murray (https://github.com/Matt711)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/19267