Fix "--executor" pytest parameter for cudf-polars (#19703)
The default value for the `--executor` pytest parameter is still `"in-memory"`. This means we are currently running the in-memory tests twice in CI.

The good news is that the `--blocksize-mode small` tests have still been hitting the `"streaming"` engine, so I'm pretty sure there is no need to worry about 25.08.

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Tom Augspurger (https://github.com/TomAugspurger)
  - Matthew Murray (https://github.com/Matt711)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/19703