Use the nvCOMP 5.0 API to better estimate decompression memory requirements (#19616)
nvCOMP 5 adds synchronous APIs that more accurately determine the scratch size requirements for decompression (ZSTD only for now). 

This PR integrates the new API in `batched_decompress`, as well as into the chunked reader (where it's used to determine memory overhead from decompression). The new API is expensive to call, as it launches a kernel. For this reason, it's used differently than the old API in the chunked reader. The result may be imprecise when blocks have very uneven compression ratio, but it should be minor given that the new API reduces the memory overhead significantly*.

Another change - the temp sizes in the chunked reader are now stored between subpasses so we don't perform redundant computation.

*measured up to 2.2x lower peak memory use in micro-benchmarks.

Did not measure any performance overhead from the use of the new API in the non-chunked read_parquet benchmarks.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - https://github.com/nvdbaranec
  - Nghia Truong (https://github.com/ttnghia)
  - Karthikeyan (https://github.com/karthikeyann)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/19616